test(data-model): split the `value-visit` tests to match the submodule
One test file per source file under `test/value-visit/`, mirroring
`src/value-visit/`, with the recording visitor and the result-form builders
shared from `Recorder.ts`. Tests import from the submodule barrel, except
for `VisitInProgress`, which the barrel deliberately omits and which is now
tested directly: three re-entry cases cover the guard against a second
top-level visit on an instance mid-visit, so that guard's coverage marker is
gone.
